extract dataclasses to help with circular import issues
